LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/back-cross-document-event-order-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: Cannot go back


Harness Error (TIMEOUT), message = null

TIMEOUT back() event ordering for cross-document traversal Test timed out
